+++ examples/complete/data.tf
@@ -11,42 +11,6 @@ data "aws_iam_policy_document" "ecs_assume_role_policy" {
   }
 }
 
-data "aws_iam_policy_document" "task_execution_role_policy_doc" {
-  #checkov:skip=CKV_AWS_111:Ensure IAM policies does not allow write access without constraints"
-  statement {
-    effect = "Allow"
-    actions = [
-      "logs:CreateLogStream",
-      "logs:PutLogEvents",
-    ]
-
-    resources = ["arn:${local.partition}:logs:::log-group:${local.name}"]
-  }
-
-  statement {
-    effect = "Allow"
-    actions = [
-      "ecr:GetAuthorizationToken",
-      "ecr:BatchCheckLayerAvailability",
-      "ecr:GetDownloadUrlForLayer",
-      "ecr:BatchGetImage",
-      "logs:CreateLogStream",
-      "logs:PutLogEvents"
-    ]
-
-    resources = ["*"]
-  }
-}
-
-data "aws_availability_zones" "available" {
-  state = "available"
-}
-data "aws_caller_identity" "current" {}
-
-data "aws_region" "current" {}
-
-data "aws_elb_service_account" "main" {}
-
 ### Access Logs Bucket Policy
 data "aws_iam_policy_document" "access_logs_bucket" {
   policy_id = "s3_bucket_lb_logs"
@@ -87,4 +51,56 @@ data "aws_iam_policy_document" "access_logs_bucket" {
       type        = "Service"
     }
   }
+}
+
+data "aws_elb_service_account" "main" {}
+
+data "aws_iam_policy_document" "task_execution_role_policy_doc" {
+  #checkov:skip=CKV_AWS_111:Ensure IAM policies does not allow write access without constraints"
+  statement {
+    effect = "Allow"
+    actions = [
+      "logs:CreateLogStream",
+      "logs:PutLogEvents",
+    ]
+
+    resources = ["arn:${local.partition}:logs:::log-group:${local.name}"]
+  }
+
+  statement {
+    effect = "Allow"
+    actions = [
+      "ecr:GetAuthorizationToken",
+      "ecr:BatchCheckLayerAvailability",
+      "ecr:GetDownloadUrlForLayer",
+      "ecr:BatchGetImage",
+      "logs:CreateLogStream",
+      "logs:PutLogEvents"
+    ]
+
+    resources = ["*"]
+  }
+}
+
+data "aws_subnets" "public" {
+  filter {
+    name   = "tag:Name"
+    values = ["${local.supporting_resources_name}*.pub.*"]
+  }
+}
+
+data "aws_vpc" "supporting" {
+  filter {
+    name   = "tag:Name"
+    values = [local.supporting_resources_name]
+  }
+}
+
+data "aws_subnet" "public" {
+  for_each = toset(data.aws_subnets.public.ids)
+  id       = each.value
+}
+
+data "aws_ecs_cluster" "ecs" {
+  cluster_name = local.supporting_resources_name
 }