+++ examples/complete/locals.tf
@@ -1,17 +1,18 @@
 locals {
-  cidr_block      = "172.16.0.0/16"
-  tag_env         = "Dev"
-  public_subnets  = cidrsubnets(local.cidr_block, 8, 8, 8)
-  azs             = flatten(data.aws_availability_zones.available.names)
-  name            = "complete-example"
-  partition       = data.aws_partition.current.partition
-  bucket          = "${local.name}-access-logs-bucket"
-  service_account = data.aws_elb_service_account.main.arn
-  tags = {
-    Name               = local.name
-    Environment        = "examples"
-    "user::CostCenter" = "terraform-registry"
-  }
+  subnet_id = [
+    for i in data.aws_subnet.public : i.id
+  ]
+
+  name                      = "complete-ecs-example"
+  public_subnets            = local.subnet_id
+  supporting_resources_name = "terraform-aws-ecs-service"
+  vpc_id                    = data.aws_vpc.supporting.id
+  vpc_cidr                  = data.aws_vpc.supporting.cidr_block
+  cluster                   = data.aws_ecs_cluster.ecs.arn
+  partition                 = data.aws_partition.current.partition
+  bucket                    = "${local.name}-access-logs-bucket"
+  service_account           = data.aws_elb_service_account.main.arn
+
   default_container_definitions = jsonencode(
     [
       {
@@ -29,4 +30,10 @@ locals {
       }
     ]
   )
+
+  tags = {
+    Name               = local.name
+    Environment        = "examples"
+    "user::CostCenter" = "terraform-registry"
+  }
 }