+++ examples/complete/main.tf
@@ -1,30 +1,3 @@
-#####
-module "vpc" {
-  source               = "git::https://github.com/boldlink/terraform-aws-vpc.git?ref=2.0.3"
-  cidr_block           = local.cidr_block
-  name                 = local.name
-  enable_dns_support   = true
-  enable_dns_hostnames = true
-  account              = data.aws_caller_identity.current.account_id
-  region               = data.aws_region.current.name
-
-  ## public Subnets
-  public_subnets          = local.public_subnets
-  availability_zones      = local.azs
-  map_public_ip_on_launch = true
-  tag_env                 = local.tag_env
-}
-
-module "kms_key" {
-  source                  = "boldlink/kms/aws"
-  description             = "A test kms key for ecs cluster"
-  create_kms_alias        = true
-  alias_name              = "alias/${local.name}"
-  enable_key_rotation     = true
-  deletion_window_in_days = 7
-  tags                    = local.tags
-}
-
 module "access_logs_bucket" {
   source        = "boldlink/s3/aws"
   version       = "2.2.0"
@@ -32,29 +5,6 @@ module "access_logs_bucket" {
   force_destroy = true
   bucket_policy = data.aws_iam_policy_document.access_logs_bucket.json
   tags          = local.tags
-}
-
-resource "aws_cloudwatch_log_group" "cluster" {
-  #checkov:skip=CKV_AWS_158:Ensure that CloudWatch Log Group is encrypted by KMS"
-  name              = "${local.name}-log-group"
-  retention_in_days = 0
-  tags              = local.tags
-}
-
-module "cluster" {
-  source = "git::https://github.com/boldlink/terraform-aws-ecs-cluster.git?ref=1.0.1"
-  name   = local.name
-  configuration = {
-    execute_command_configuration = {
-      kms_key_id = module.kms_key.key_id
-      log_configuration = {
-        cloud_watch_encryption_enabled = true
-        cloud_watch_log_group_name     = aws_cloudwatch_log_group.cluster.name
-        s3_bucket_encryption_enabled   = false
-      }
-      logging = "OVERRIDE"
-    }
-  }
 }
 
 module "ecs_service_lb" {
@@ -66,12 +16,12 @@ module "ecs_service_lb" {
   name                     = "${local.name}-service"
   family                   = "${local.name}-task-definition"
   network_configuration = {
-    subnets          = flatten(module.vpc.public_subnet_id)
+    subnets          = local.public_subnets
     assign_public_ip = true
   }
-  alb_subnets                = flatten(module.vpc.public_subnet_id)
-  cluster                    = module.cluster.id
-  vpc_id                     = module.vpc.id
+  alb_subnets                = local.public_subnets
+  cluster                    = local.cluster
+  vpc_id                     = local.vpc_id
   task_role_policy           = data.aws_iam_policy_document.ecs_assume_role_policy.json
   task_execution_role        = data.aws_iam_policy_document.ecs_assume_role_policy.json
   task_execution_role_policy = data.aws_iam_policy_document.task_execution_role_policy_doc.json
@@ -115,7 +65,7 @@ module "ecs_service_lb" {
       from_port   = 80
       to_port     = 80
       protocol    = "tcp"
-      cidr_blocks = [local.cidr_block]
+      cidr_blocks = [local.vpc_cidr]
     }
   ]