+++ examples/minimum/data.tf
@@ -1,0 +1,3 @@
+data "aws_ecs_cluster" "ecs" {
+  cluster_name = local.supporting_resources_name
+}